pyproject.toml
[project]
name = "semanticist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.0",
    "diffusers[torch]>=0.35.1",
    "einops>=0.8.1",
    "huggingface-hub>=0.34.4",
    "lightning>=2.5.5",
    "matplotlib>=3.10.5",
    "numpy==1.26.4",
    "omegaconf>=2.3.0",
    "opencv-python-headless>=4.11.0.86",
    "pycocotools>=2.0.10",
    "safetensors>=0.6.2",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "submitit>=1.5.3",
    "tensorboard>=2.20.0",
    "timm>=1.0.19",
    "torch-fidelity",
    "torchmetrics>=1.8.1",
    "transformers>=4.55.4",
    "xformers>=0.0.32.post2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["semanticist"]

[tool.uv.sources]
torch-fidelity = { git = "https://github.com/LTH14/torch-fidelity.git", rev = "master" }
timm = { git = "https://github.com/rwightman/pytorch-image-models.git" }
